# Optimize TRY(CAST) for many failed rows