--- a/velox/expression/TryExpr.cpp
+++ b/velox/expression/TryExpr.cpp
@@ -37,6 +37,11 @@ void TryExpr::evalSpecialForm(
   // parent TRY expression, so the parent won't incorrectly null out rows that
   // threw exceptions which this expression already handled.
   ScopedVarSetter<ErrorVectorPtr> errorsSetter(context.errorsPtr(), nullptr);
+
+  // Allocate error vector to avoid repeated re-allocations for every failed
+  // row.
+  context.ensureErrorsVectorSize(rows.end());
+
   inputs_[0]->eval(rows, context, result);
 
   nullOutErrors(rows, context, result);
@@ -61,6 +66,11 @@ void TryExpr::evalSpecialFormSimplified(
   // parent TRY expression, so the parent won't incorrectly null out rows that
   // threw exceptions which this expression already handled.
   ScopedVarSetter<ErrorVectorPtr> errorsSetter(context.errorsPtr(), nullptr);
+
+  // Allocate error vector to avoid repeated re-allocations for every failed
+  // row.
+  context.ensureErrorsVectorSize(rows.end());
+
   inputs_[0]->evalSimplified(rows, context, result);
 
   nullOutErrors(rows, context, result);
@@ -134,9 +144,10 @@ void TryExpr::nullOutErrors(
       }
     } else {
       if (result.unique() && result->isNullsWritable()) {
+        auto* rawNulls = result->mutableRawNulls();
         rows.applyToSelected([&](auto row) {
           if (row < errors->size() && !errors->isNullAt(row)) {
-            result->setNull(row, true);
+            bits::setNull(rawNulls, row, true);
           }
         });
       } else {