Summary:
TRY(CAST(...)) is up to 5x slower than TRY_CAST when many rows fail.

The profile reveals that 30% of cpu time goes to
EvalCtx::ensureErrorsVectorSize. For every row that fails, we call
EvalCtx::ensureErrorsVectorSize to resize the error vector to accommodate that
row. When many rows fail we end up resizing a lot: resize(1), resize(2), resize
(3),....resize(n). Fix this by pre-allocating error vector in TryExpr.

Another 30% of cpu time goes into managing reference counts for
`std::shared_ptr<std::exception_ptr>` stores in the errors vector. We do not
need error details for TRY, hence, no need to store these values. Fix this by
making error values in error vector optional and updating EvalCtx::setStatus to
not write these under TRY.

Finally, replace BaseVector::setNull with bits::setNull in TryExpr::nullOutErrors.

{F1647570993}

Combined these optimizations improve try(cast) benchmark by 3x.

Before:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##try_cast_invalid_empty_input                          2.41ms    415.65
cast##tryexpr_cast_invalid_empty_input                     24.90ms     40.16
cast##try_cast_invalid_nan                                  5.76ms    173.52
cast##tryexpr_cast_invalid_nan                             27.75ms     36.03
```

After:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##try_cast_invalid_empty_input                          2.27ms    439.78
cast##tryexpr_cast_invalid_empty_input                      5.93ms    168.61
cast##try_cast_invalid_nan                                  5.64ms    177.35
cast##tryexpr_cast_invalid_nan                              9.44ms    105.94
```

Differential Revision: D57704232